% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCanopyVolume.R
\name{calculateCanopyVolume}
\alias{calculateCanopyVolume}
\title{Calculate Canopy Volume}
\usage{
calculateCanopyVolume(lazFiles, shapefile, threshold, outputDir)
}
\arguments{
\item{lazFiles}{filenames(list of strings): Set the working directory to where the LAZ files are located and access a list of LAZ files}

\item{shapefile}{filename(string): Full path to the initial shapes shapefile matching the UNIQUE_ID with the LAZ files, and it must contain ELEVATION and HEIGHT data}

\item{threshold}{numeric: The height at which to clip all LAZ points below it out.}

\item{outputDir}{directory(string): Full path to the output location.}
}
\value{
Clipped LAZ and an updated shapefile containing the CANOPYVOL field
}
\description{
Calculate Canopy Volume
}
\examples{
lazDir <- readClipboard() # Copy the path to the LAZ files from File Explorer
setwd(lazDir)
lazFiles <- list.files(pattern="*.laz")
shapefile <- "initialShapes.shp"
dir.create("output")
calculateCanopyVolume(lazFiles,shapefile,threshold=10,outputDir=paste0(getwd(),"//output"))
}
