% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeZgeometry_fromHeight.R
\name{changeZgeometry_fromHeight}
\alias{changeZgeometry_fromHeight}
\title{Change Z Geometry for a shapefile based on HEIGHT column
Uses the numerical values in the HEIGHT column to add to its present elevation (Z) data to present polygons that are Z + HEIGHT in elevation.}
\usage{
changeZgeometry_fromHeight(
  shapefile,
  output = gsub(".shp", "_fixedZ.shp", shapefile)
)
}
\arguments{
\item{shapefile}{filename : The input shapefile that you would like to increase the Z geometry for, based on the HEIGHT column. This shapefile must contain a HEIGHT column.}

\item{output}{filename : The output location for the "fixed" shapefile.}
}
\value{
shapefile
}
\description{
Change Z Geometry for a shapefile based on HEIGHT column
Uses the numerical values in the HEIGHT column to add to its present elevation (Z) data to present polygons that are Z + HEIGHT in elevation.
}
\examples{
shpDir <- readClipboard() # Read the directory path of the folder containing the shapefile you would like to fix
setwd(shpDir)
files <- list.files(pattern="*.shp")
changeZgeometry_fromHeight(files[1]) # Fixes the Z geometry of the first shapefile listed in that directory
}
