% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityGraphProduction.R
\name{intensityGraphProduction}
\alias{intensityGraphProduction}
\title{Intensity Graph Production}
\usage{
intensityGraphProduction(
  agtLAS,
  agtSHP,
  outDir,
  aoiName = "AGT",
  decid = c("AL", "AW", "PB", "BW", "SY", "BA", "BE", "MA", "EL", "AH", "HO", "WO", "OK",
    "BC", "MB", "HB", "HL", "SM", "DC", "AC", "AT", "EP", "HW"),
  conifer = c("FB", "LT", "PL", "SB", "SW", "SP", "CR", "WP", "PI", "CW", "FD", "SX"),
  dead = c("DP", "DB", "DS", "DF", "SN")
)
}
\arguments{
\item{agtLAS}{string. Location of merged AGT LAS/LAZ file. Right click on the file name while holding down the "SHIFT" key. Select "Copy as path". In the R console, type "readClipboard()" and manually manipulate the directory so that it is a string between two quotation marks without the final backslash.}

\item{agtSHP}{string. Location of merged AGT shapefile. Please follow the same instructions as agtLAS for inputting agtSHP.}

\item{outDir}{string. Output directory to store raw intensity data, raw histogram data, species table, and plots. After creating this output directory, select the file path at the top of the File Explorer window and simply type into the console - outDir <- readClipboard()}

\item{aoiName}{string. The name of the project AOI. This will be included as a title in all of the plots.}

\item{decid}{string list. List of deciduous species (NOTE : Please check the default list and if any species is not included in it, or incorrectly labelled deciduous, you will need to manually set this list).}

\item{conifer}{string list. List of coniferous species (NOTE : Please check the default list and if any species is not included in it, or incorrectly labelled conifer, you will need to manually set this list).}

\item{dead}{string list. List of dead species (NOTE : Please check the default list and if any species is not included in it, or incorrectly labelled dead, you will need to manually set this list).}
}
\value{
Four plots displaying the intensity distribution of all species within the AOI
}
\description{
This tool uses the merged LAS and merged SHP to produce an overall intensity graph. You will need an output directory which will store all  raw intensity files along with plots and the final tables indicating the number points and shapes per species and the intensity histogram data. (Option for future : it would be easy to include an option for merging results between different project areas and projections, if need be)
}
\examples{
# Copy the agtLAS path to clipboard
readClipboard()
#[1] "\"C:\\Users\\...\\2446_WF_AreaA_AGT_20230915_merged.las\""
agtLAS <- "C:\\\\Users\\\\...\\\\2446_WF_AreaA_AGT_20230915_merged.las"
# Copy agtSHP path to clipboard
readClipboard()
#[1] "\"C:\\Users\...\\2446_WF_AreaA_AGT_20230915_mergedInitialShapes.shp\""
agtSHP <- "C:\\\\Users\\\\...\\\\2446_WF_AreaA_AGT_20230915_mergedInitialShapes.shp"
outDir <- "C:\\\\IntensityGraphs\\\\WF_AreaA"
aoiName <- "West Frasier Area A"
#intensityGraphProduction(agtLAS,agtSHP,outDir,aoiName)
}
