% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinningPlots.R
\name{pinningPlots}
\alias{pinningPlots}
\title{Pinning Plots}
\usage{
pinningPlots(
  parentDirectory,
  folderCount,
  overstory = "CN",
  understory = "CU",
  windowSize = 2,
  minHeight = 5
)
}
\arguments{
\item{parentDirectory}{string: The parent directory of your InProgress folder. The final folder in this directory must be your name (example:"...GT_Boxes/_inProgress/Michael). The easiest way to use this is to copy the directory (in Windows File Explorer, copy the entire directory at the top of the window, and in your R window, assign it to a variable like this : parentDir <- readClipboard(). Then use the 'parentDir' variable in this place).}

\item{folderCount}{numerical: Since you will be working one folder at a time, start with the number 1 and after you have finished pinning that plot, use the number 2 (or if you immediately move that first plot to the Final folder, you can use 1 again).}

\item{overstory}{string: Overstory tree type. Default is 'CN', but you can change to 'DC' after viewing the plot.}

\item{understory}{string: Understory tree type. Default is 'CU', but you can change to 'DU' after viewing the plot.}

\item{windowSize}{numeric or function: Length or diameter of the moving window used to detect the local maxima in the units of the input data (usually meters). If it is numeric a fixed window size is used. If it is a function, the function determines the size of the window at any given location on the canopy. By default function takes the height of a given pixel or point as its only argument and return the desired size of the search window when centered on that pixel/point. This can be controled with the 'ws_args' parameter}

\item{minHeight}{numeric: Minimum height of a tree. Threshold below which a pixel or a point cannot be a local maxima. Default is 2.}
}
\value{
A roughly pinned plot in a treeTops.shp file. Must of the understory will be excluded.
}
\description{
This function will help build a foundation for pinning your plots. It is not perfect, and it will not capture any of the understory, but it will help to get you started.
}
\examples{
# Use the following template
# parentDir <- readClipboard() # after copying the file directory to your clipboard
# pinningPlots(parentDir,1,'Michael')
}
