% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassifyLAZ.R
\name{reclassifyLAZ}
\alias{reclassifyLAZ}
\title{Reclassify LAZ to LASUNCLASSIFIED}
\usage{
reclassifyLAZ(
  lazFile,
  lazClasses = c(LASLOWVEGETATION, LASMEDIUMVEGETATION, LASHIGHVEGETATION, LASBUILDING)
)
}
\arguments{
\item{lazFile}{filename : This can be an item in a list or a singular filename}

\item{lazClasses}{list : This is a list of classes to be reclassified to LASUNCLASSIFIED. Please refer to the format in which they are already classed}
}
\value{
LAZ/LAS files
}
\description{
This tool reclassifies all of the LAZ/LAS files within a list to change a series of classes to become unclassified. This is useful for seg cal and other various PowerPoint presentations.
}
\examples{
parentDirectory <- readClipboard() # Read the directory of a folder containing various LAS/LAZ files into your clipboard and run this line
setwd(parentDirectory)
files <- list.files(pattern="*.laz")
for(file in files){
  reclassifyLAZ(file)
}
}
