% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewAllLAZ.R
\name{viewAllLAZ}
\alias{viewAllLAZ}
\title{viewAllLAZ
Displays all LAZ files  from a list, one by one.}
\usage{
viewAllLAZ(lazFile, plotView = FALSE)
}
\arguments{
\item{lazFile}{filenames: A string representative of a singular LAZ/LAS file}

\item{plotView}{boolean: TRUE if you want a generalized view of the point cloud with file name, FALSE if you want a superior rendering with poor filename quality}
}
\value{
The RGL window displays a LAS point cloud
}
\description{
viewAllLAZ
Displays all LAZ files  from a list, one by one.
}
\examples{
parentDir <- readClipboard() # Read the directory path of a folder containing several LAZ/LAS files
setwd(parentDir)
files <- list.files(pattern="*.laz")
for(file in files){viewAllLAS(file)}
}
