% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelSegCal_DisplayConfigLog.R
\name{voxelSegCal_DisplayConfigLog}
\alias{voxelSegCal_DisplayConfigLog}
\title{Display Configuration Logs - Voxel Segmentation Calibration}
\usage{
voxelSegCal_DisplayConfigLog(resultsDirectory)
}
\arguments{
\item{resultsDirectory}{directory(string) : Directory containing all of the folders for the results}
}
\value{
segParameters data frame containing all of the logs
}
\description{
Display Configuration Logs - Voxel Segmentation Calibration
}
\examples{
#voxelSegCal_DisplayConfigLog("C:\\Tolko\\Greensled\\BaseTarget")
}
