% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelSegCal_NewSegmentation.R
\name{voxelSegCal_NewSegmentation}
\alias{voxelSegCal_NewSegmentation}
\title{New Segmentation - Voxel Segmentation Calibration}
\usage{
voxelSegCal_NewSegmentation(
  CFG_filename,
  resultsDirectory,
  voxelSegmentationEXE_directory,
  Dropbox_directory,
  allPinTestSummary_directory
)
}
\arguments{
\item{CFG_filename}{filename(string) : Full path to the input CFG file to run the segmentation from}

\item{resultsDirectory}{directory(string) : Directory containing all of the folders for the results}

\item{voxelSegmentationEXE_directory}{filename(string) : Full path to the VoxelSegmentationCPU.exe with all helper files in the same directory}

\item{Dropbox_directory}{directory(string) : Directory on Dropbox where to copy the new target folder to}

\item{allPinTestSummary_directory}{filename(string) :  Full path to the CSV containing the statistical results based on POIs}
}
\value{
New voxel segmentation with statistical results from the main columns
}
\description{
New Segmentation - Voxel Segmentation Calibration
}
\examples{
#cfgFilename <- "C:\\Forsite\\Tolko\\GreenSled_voxelSeg\\_TolkoGS_NewFormat_Voxel_SegCal.cfg"
#resultsDirectory <- "C:\\Forsite\\Tolko\\GreenSled_voxelSeg\\BaseTarget"
#voxelSegEXE_dir <- "C:\\Forsite\\voxelSeg_training\\VoxelSegEXE\\VoxelSegmentationCPU.exe"
#dropbox_dir <- "C:\\Users\\MichaelBurnett\\LiDAR Inventory group Dropbox\\Project Data\\OR\\Tolko\\Tolko_Sask\\_GreenSled\\SegCal\\_Results"
#allPinTestSummary_dir <- "C:\\Forsite\\Tolko\\GreenSled_voxelSeg\\BaseTarget\\all_Pin_All_Test_Summary_No_Edge.csv"
#voxelSegCal_NewSegmentation(cfgFilename,resultsDirectory,voxelSegEXE_dir,allPinTestSummary_dir)

}
