% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelSegCal_sourceIDviewing.R
\name{voxelSegCal_sourceIDviewing}
\alias{voxelSegCal_sourceIDviewing}
\title{Source ID Viewing for Voxel Segmentation LAS}
\usage{
voxelSegCal_sourceIDviewing(resultsDirectory, targetDir, plot_prefix, plotNUM)
}
\arguments{
\item{resultsDirectory}{directory : The full directory to the location of the RESULTS from the voxel segmentation (The parent directory of the target segmentation folders)}

\item{targetDir}{string : The folder name for the segmentation results to be displayed}

\item{plot_prefix}{string : The prefix to each plot. Please include every character up until the number for the plot (for example: "Tolko_Greensled_Plot_"). The "_DISPLAY.laz" is implied.}

\item{plotNUM}{numerical : The plot number to be displayed}
}
\value{
An RGL window displaying the classified point cloud for the plot.
}
\description{
Source ID Viewing for Voxel Segmentation LAS
}
\examples{
resultsDirectory <- "C:\\\\Forsite\\\\Tolko\\\\GreenSled_voxelSeg\\\\BaseTarget"
targetDir <- "target36"
plot_prefix <- "Tolko_Greensled_Plot_"
plotNUM <- 100

voxelSegCal_sourceIDviewing(resultsDirectory,targetDir,plot_prefix, plotNUM)
}
