% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewAllLAZ.R
\name{viewAllLAZ}
\alias{viewAllLAZ}
\title{viewAllLAZ
Displays all LAZ files  from a list, one by one.}
\usage{
viewAllLAZ(lazFile)
}
\arguments{
\item{lazFile}{filename: A string representative of a singular LAZ/LAS file}
}
\value{
The RGL window displays a LAS point cloud
}
\description{
viewAllLAZ
Displays all LAZ files  from a list, one by one.
}
\examples{
parentDir <- readClipboard() # Read the directory path of a folder containing several LAZ/LAS files
setwd(parentDir)
files <- list.files(pattern="*.laz")
for(file in files){
  viewAllLAZ(file)
}
}
